requirements.txt
asgiref==3.5.2
autopep8==1.6.0
click==8.1.3
colorama==0.4.5
crispy-bootstrap5==0.6
Django==4.0.5
django-crispy-forms==1.14.0
djlint==1.3.0
et-xmlfile==1.1.0
html-tag-names==0.1.2
html-void-elements==0.1.0
importlib-metadata==4.12.0
mysql==0.0.3
mysqlclient==2.1.0
numpy==1.22.0
openpyxl==3.0.10
pandas==1.4.2
pathspec==0.9.0
Pillow==9.2.0
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.6.2
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
zipp==3.8.0
